More Python compatibility
