+++ numba/cuda/cudadrv/devicearray.py
@@ -8,6 +8,7 @@ from __future__ import print_function, absolute_import, division
 import warnings
 import math
 import copy
+import six
 from ctypes import c_void_p
 
 import numpy as np
@@ -16,11 +17,6 @@ from . import driver as _driver
 from . import devices
 from numba import dummyarray, types, numpy_support
 from numba.unsafe.ndarray import to_fixed_tuple
-
-try:
-    long
-except NameError:
-    long = int
 
 
 def is_cuda_ndarray(obj):
@@ -41,7 +37,7 @@ def verify_cuda_ndarray_interface(obj):
     requires_attr('shape', tuple)
     requires_attr('strides', tuple)
     requires_attr('dtype', np.dtype)
-    requires_attr('size', (int, long))
+    requires_attr('size', six.integer_types)
 
 
 def require_cuda_ndarray(obj):
@@ -75,9 +71,9 @@ class DeviceNDArrayBase(object):
         gpu_data
             user provided device memory for the ndarray data buffer
         """
-        if isinstance(shape, (int, long)):
+        if isinstance(shape, six.integer_types):
             shape = (shape,)
-        if isinstance(strides, (int, long)):
+        if isinstance(strides, six.integer_types):
             strides = (strides,)
         self.ndim = len(shape)
         if len(strides) != self.ndim:
