+++ numba/cuda/tests/cudapy/test_forall.py
@@ -6,7 +6,6 @@ from numba import cuda
 import numba.unittest_support as unittest
 from numba.cuda.testing import skip_on_cudasim, SerialMixin
 
-@skip_on_cudasim('forall API unsupported in the simulator')
 class TestForAll(SerialMixin, unittest.TestCase):
     def test_forall_1(self):
